Setup settings for different stages
